fix: Make TaskCard and TaskControlPanel components
