+++ src/app/(admin)/tasks/page.tsx
@@ -1,11 +1,12 @@
 "use client";
 
-import { ControlFilterDropdown, ControlSearch, ControlStatusPills } from "@/components/ControlPanel";
-import { TaskSchema } from "@/components/data-table/table-widget-defs";
 import { useState } from "react";
+import { TaskSchema } from "@/components/data-table/table-widget-defs";
+import { TaskCard } from "@/components/TaskCard";
+import { TasksControlPanel } from "@/components/TasksControlPanel";
 
 export default function Tasks() {
-  //TODO: integrate backend instead of using mock data
+  // TODO: integrate backend instead of using mock data
   const tasks: TaskSchema[] = [
     {
       id: 1,
@@ -122,14 +123,16 @@ export default function Tasks() {
   const [status, setStatus] = useState("All");
   const [surveys, setSurveys] = useState("All Tasks");
 
+  const filteredTasks = filterTasks(tasks, status, surveys);
+
   return (
     <main className="flex-1 min-w-0 bg-[#f6f2ec]">
-      <div className="flex flex-col ml-[20px] mt-[15px] mr-[20px] gap-[20px]">
+      <div className="flex flex-col min-w-0 ml-[20px] mt-[15px] mr-[20px] gap-[20px]">
         <header className="flex items-center justify-between pt-5">
           <h1 className="flex text-[36px] font-[Constantia] font-bold">Tasks</h1>
         </header>
 
-        <div className="rounded-xl border-[1px] border-[#d8d3ca] drop-shadow-sm bg-[#ebe7de] w-auto h-auto">
+        <div className="rounded-xl border-[1px] border-[#d8d3ca] drop-shadow-sm bg-[#ebe7de] w-full min-w-0">
           <TasksControlPanel
             setStatusFunction={setStatus}
             setSurveyFunction={setSurveys}
@@ -140,14 +143,14 @@ export default function Tasks() {
         <div className="flex flex-col rounded-xl border-[1px] border-[#d8d3ca] drop-shadow-sm bg-[#ebe7de] w-auto h-[570px] overflow-hidden">
           <div className="flex-1 overflow-y-scroll no-scrollbar p-[15px]">
             <div className="flex flex-col gap-[10px]">
-              {filter(tasks, status, surveys).map((task) => {
-                return <TaskCard key={String(task.id)} task={task} />;
-              })}
+              {filteredTasks.map((task) => (
+                <TaskCard key={String(task.id)} task={task} />
+              ))}
             </div>
           </div>
 
           <div className="border-t border-[#ded9cf] border-t-[1.5px] px-[15px] py-[10px] text-[#6b6661] text-[14px] bg-[#ebe7de] font-semibold">
-            Showing {filter(tasks, status, surveys).length} tasks
+            Showing {filteredTasks.length} tasks
           </div>
         </div>
       </div>
@@ -155,113 +158,11 @@ export default function Tasks() {
   );
 }
 
-interface TaskControlPanelProps {
-  setStatusFunction: (status: string) => void;
-  setSurveyFunction: (survey: string) => void;
-  searchFunction: (query: string) => void;
-}
-
-//TODO: integrate filterability
-function TasksControlPanel(props: TaskControlPanelProps) {
-  const CONTROL_STATUS_OPTIONS = ["All", "Done"];
-  const ASSIGNEE_STATUS_OPTIONS = ["All Assignees"];
-  const SURVEY_OPTIONS = ["All Tasks", "Surveys Needed", "Surveys Complete"];
-  const [searchQuery, setSearchQuery] = useState("");
-  const [statusActiveIndex, setStatusActiveIndex] = useState(0);
-  const [assigneesActiveIndex, setAssigneesActiveIndex] = useState(0);
-  const [surveysActiveIndex, setSurveyActiveIndex] = useState(0);
-  const QUERY_DELAY = 0;
-
-  return (
-    <div className="w-full">
-      <div className="flex w-full flex-col justify-center rounded-[24px] sm:rounded-[32px] bg-inherit px-4 sm:px-6 lg:px-10 py-6 sm:py-8 gap-6">
-        <ControlSearch
-          query={searchQuery}
-          onQueryChange={setSearchQuery}
-          searchDelay={QUERY_DELAY}
-          searchFunction={(query: string) => {
-            const trimmedQuery = query.trimStart();
-            props.searchFunction(trimmedQuery);
-          }}
-          placeholder={"Search tasks, messages, or assignees..."}
-        />
-        <div className="w-full">
-          <div className="flex flex-col gap-4 w-full">
-            <div className="flex flex-col gap-4 w-full xl:flex-row xl:items-end">
-              <div className="w-full xl:w-auto">
-                <ControlStatusPills
-                  pillClassName="w-full sm:w-auto sm:min-w-[220px] lg:min-w-[256px]"
-                  activeIndex={statusActiveIndex}
-                  onActiveIndexChange={setStatusActiveIndex}
-                  text="Status"
-                  options={CONTROL_STATUS_OPTIONS}
-                  delay={QUERY_DELAY}
-                  delayFunction={(status: string) => props.setStatusFunction(status)}
-                />
-              </div>
-              <div className="flex flex-col gap-4 w-full sm:flex-row">
-                <ControlFilterDropdown
-                  triggerClassName="w-full min-w-0 sm:w-[240px] lg:w-[456px]"
-                  label="Assignee"
-                  activeIndex={assigneesActiveIndex}
-                  onActiveIndexChange={setAssigneesActiveIndex}
-                  dropDown={ASSIGNEE_STATUS_OPTIONS}
-                  delay={QUERY_DELAY}
-                  delayFunction={() =>
-                    //TODO: once backend is implemented, implement assignee filter logic
-                    console.log("PlaceHolder")
-                  }
-                />
-                <ControlFilterDropdown
-                  triggerClassName="w-full min-w-0 sm:w-[240px] lg:w-[456px]"
-                  label="Surveys"
-                  activeIndex={surveysActiveIndex}
-                  onActiveIndexChange={setSurveyActiveIndex}
-                  dropDown={SURVEY_OPTIONS}
-                  delay={QUERY_DELAY}
-                  delayFunction={(status: string) => props.setSurveyFunction(status)}
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function TaskCard(props: { task: TaskSchema }) {
-  const task = props.task;
-  return (
-    <div className="flex flex-col rounded-xl bg-[#ffffff] w-auto h-fit p-[16px]">
-      <div className="flex flex-col gap-[10px] mb-[15px]">
-        <p className="text-black text-[16px] font-[Constantia] font-semibold">{task.title}</p>
-        <p className="text-[#6b6661] text-[16px] font-[Constantia]">{task.message}</p>
-      </div>
-      <hr className="border-[.5px] border-[#e8e6e0]"></hr>
-      <div className="flex flex-row gap-[8px]">
-        <div className="flex flex-row gap-[5px]">
-          <p className="font-semibold">{task.assignees?.length} </p>
-          <p className="text-[#6b6661]">assignees</p>
-        </div>
-        <div className="flex flex-row">
-          <p className="font-semibold">{task.is_complete ? task.surveys_needed : 0}</p>
-          <p className="text-[#6b6661]">/{task.surveys_needed}</p>
-        </div>
-        <div className="flex flex-row">
-          <p className="text-[#6b6661]">Due {task.completion_date?.substring(0, 11)}</p>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function filter(tasks: TaskSchema[], status: string, surveys: string) {
-  tasks = tasks.filter((task) => {
+function filterTasks(tasks: TaskSchema[], status: string, surveys: string) {
+  return tasks.filter((task) => {
     return (
-      (status == "All" || task.is_complete) &&
-      (surveys == "All Tasks" || Number(surveys == "Surveys Needed") ^ Number(task.is_complete))
+      (status === "All" || task.is_complete) &&
+      (surveys === "All Tasks" || Number(surveys === "Surveys Needed") ^ Number(task.is_complete))
     );
   });
-  return tasks;
 }
